+++ packages/contracts/test/registry/ens-subdomain-registry.ts
@@ -9,7 +9,7 @@ import {
   ENSRegistry,
 } from '../../typechain';
 import {customError} from '../test-utils/custom-error-helper';
-import {ensDomainHash, ensLabelHash} from '../../utils/ensHelpers';
+import {ensDomainHash, ensLabelHash} from '../../utils/ens';
 
 import {setupResolver} from '../test-utils/ens';
